docs: improve code comments, fix typos, move a few functions
