+++ lib/actions/field-trip.js
@@ -3,7 +3,10 @@ import {
   getRoutingParams,
   planParamsToQueryAsync
 } from '@opentripplanner/core-utils/lib/query'
-import { OTP_API_TIME_FORMAT } from '@opentripplanner/core-utils/lib/time'
+import {
+  OTP_API_DATE_FORMAT,
+  OTP_API_TIME_FORMAT
+} from '@opentripplanner/core-utils/lib/time'
 import { randId } from '@opentripplanner/core-utils/lib/storage'
 import moment from 'moment'
 import { serialize } from 'object-to-formdata'
@@ -11,8 +14,6 @@ import qs from 'qs'
 import { createAction } from 'redux-actions'
 
 import {
-  getFieldTripGroupCapacityForMode,
-  getFormattedRequestTravelDate,
   getGroupSize,
   getTripFromRequest,
   lzwEncode,
@@ -164,6 +165,14 @@ export function editSubmitterNotes (request, submitterNotes) {
   }
 }
 
+/**
+ * Validates and saves the currently active field trip itineraries to the
+ * appropriate inbound or outbound trip of a field trip request.
+ *
+ * @param  {Object} request  The field trip request
+ * @param  {boolean} outbound If true, save the current itineraries to the
+ *                            outbound field trip journey.
+ */
 export function saveRequestTripItineraries (request, outbound) {
   return async function (dispatch, getState) {
     const state = getState()
@@ -204,6 +213,7 @@ export function saveRequestTripItineraries (request, outbound) {
 /**
  * Checks that a group plan is valid for a given request, i.e., that it occurs
  * on the requested travel date.
+ *
  * @param  request   field trip request
  * @param  itineraries the currently active itineraries
  * @return true if invalid
@@ -233,11 +243,13 @@ function fieldTripPlanIsInvalid (request, itineraries) {
     return true
   }
 
-  // FIXME More checks? E.g., origin/destination
-
   return false
 }
 
+/**
+ * Returns the earliest start time (in unix epoch milliseconds) in the given
+ * itineraries.
+ */
 function getEarliestStartTime (itineraries) {
   return itineraries.reduce(
     (earliestStartTime, itinerary) =>
@@ -246,6 +258,14 @@ function getEarliestStartTime (itineraries) {
   )
 }
 
+/**
+ * Returns true if the user confirms the overwrite of an existing set of
+ * itineraries planned for the inbound or outbound part of the field trip.
+ *
+ * @param  {Object} request  The field trip request
+ * @param  {boolean} outbound If true, save the current itineraries to the
+ *                            outbound field trip journey.
+ */
 function overwriteExistingRequestTripsConfirmed (request, outbound) {
   const type = outbound ? 'outbound' : 'inbound'
   const preExistingTrip = getTripFromRequest(request, outbound)
@@ -257,7 +277,8 @@ function overwriteExistingRequestTripsConfirmed (request, outbound) {
 }
 
 /**
- * Construct data for request to save
+ * Constructs data used to save a set of itineraries for the inbound or outbound
+ * part of the field trip.
  */
 function makeSaveFieldTripItinerariesData (request, outbound, state) {
   const { fieldTrip, session } = state.callTaker
@@ -266,6 +287,7 @@ function makeSaveFieldTripItinerariesData (request, outbound, state) {
   const fieldTripModuleConfig = getModuleConfig(state, Modules.FIELD_TRIP)
   const itineraries = getActiveItineraries(state)
 
+  // initialize base object
   const data = {
     gtfsTrips: [],
     itins: [],
@@ -283,6 +305,8 @@ function makeSaveFieldTripItinerariesData (request, outbound, state) {
     }
   }
 
+  // itierate through itineraries to construct itinerary and gtfsTrip data to
+  // save the itineraries to otp-datastore.
   itineraries.forEach((itinerary, itinIdx) => {
     const itineraryDataToSave = {
       itinData: lzwEncode(JSON.stringify(itinerary)),
@@ -323,6 +347,9 @@ function makeSaveFieldTripItinerariesData (request, outbound, state) {
   return serialize(data, { indices: true })
 }
 
+/**
+ * Creates an OTP-style location string based on the location name, lat and lon.
+ */
 function getOTPLocationString (location) {
   const llString = `${location.lat},${location.lon}`
   return location.name
@@ -330,6 +357,10 @@ function getOTPLocationString (location) {
     : llString
 }
 
+/**
+ * Begins the process of making trip requests to find suitable itineraries for
+ * either an inbound or outbound journey of a field trip.
+ */
 export function planTrip (request, outbound) {
   return function (dispatch, getState) {
     dispatch(setGroupSize(getGroupSize(request)))
@@ -339,6 +370,10 @@ export function planTrip (request, outbound) {
   }
 }
 
+/**
+ * Sets the appropriate request parameters for an outbound journey of a field
+ * trip.
+ */
 function planOutbound (request) {
   return async function (dispatch, getState) {
     const {config} = getState().otp
@@ -358,6 +393,10 @@ function planOutbound (request) {
   }
 }
 
+/**
+ * Sets the appropriate request parameters for an inbound journey of a field
+ * trip.
+ */
 function planInbound (request) {
   return async function (dispatch, getState) {
     const {config} = getState().otp
@@ -379,7 +418,7 @@ function planInbound (request) {
 
 /**
  * Makes appropriate OTP requests until enough itineraries have been found to
- * accomodate the field trip group.
+ * accommodate the field trip group.
  */
 function makeFieldTripPlanRequests (request) {
   return async function (dispatch, getState) {
@@ -415,7 +454,7 @@ function makeFieldTripPlanRequests (request) {
 
     let shouldContinueSearching = true
 
-    // make requests until enough itineraries have been found to accomodate
+    // make requests until enough itineraries have been found to accommodate
     // group
     while (shouldContinueSearching) {
       numRequests++
@@ -472,10 +511,17 @@ function makeFieldTripPlanRequests (request) {
   }
 }
 
+/**
+ * Makes a request to get data about other planned field trips happening on a
+ * particular date.
+ */
 async function getTripIdsForTravelDate (dispatch, state, request) {
   const {datastoreUrl} = state.otp.config
   const {sessionId} = state.callTaker.session
-  const formattedTravelDate = getFormattedRequestTravelDate(request, 'MM/DD/YYYY')
+  const formattedTravelDate = getFormattedRequestTravelDate(
+    request,
+    FIELD_TRIP_DATE_FORMAT
+  )
   const params = {
     date: formattedTravelDate,
     sessionId
@@ -503,6 +549,11 @@ async function getTripIdsForTravelDate (dispatch, state, request) {
   await dispatch(receivedTravelDateTrips(trips))
 }
 
+/**
+ * Makes the needed requests to the OTP index API to obtain semantic trip hashes
+ * for any tripIds in the active itineraries that haven't already been obtained
+ * from the OTP index API.
+ */
 function getTripHashesFromActiveItineraries () {
   return function (dispatch, getState) {
     const state = getState()
@@ -528,6 +579,22 @@ function getTripHashesFromActiveItineraries () {
   }
 }
 
+/**
+ * Analyzes the current itineraries from each response in the active search to
+ * appropriately assign the field trip group to subgroups that use each
+ * itinerary according to the capacity avaiable in each itinerary.
+ *
+ * @return {Object} result
+ * @return {Object} result.assignedItinerariesByResponse An Object organized by
+ *  response index and list of itineraries that can be used to update the state
+ *  with modified itinerary objects that can help display the field trip group
+ *  size assigned to each itinerary.
+ * @return {Object} result.remainingGroupSize The remaining about of people in
+ *  the field trip group that have yet to be assigned to an itinerary.
+ * @return {Object} result.tripsToBanInSubsequentSearches An array of strings
+ *  representing tripIds that should be added to the list of tripIds to ban when
+ *  making requests for additional itineraries from OTP.
+ */
 function checkValidityAndCapacity (state, request) {
   const fieldTripModuleConfig = getModuleConfig(state, Modules.FIELD_TRIP)
   const minimumAllowableRemainingCapacity =
@@ -620,7 +687,7 @@ function checkValidityAndCapacity (state, request) {
 
       if (itineraryCapacity > 0) {
         // itinerary is possible, add to list and update remaining group size.
-        // A field trip response is gauranteed to have only one itinerary, so it
+        // A field trip response is guaranteed to have only one itinerary, so it
         // ok to set the itinerary by response as an array with a single
         // itinerary.
         assignedItinerariesByResponse[responseIdx] = [{
@@ -639,6 +706,48 @@ function checkValidityAndCapacity (state, request) {
   }
 }
 
+/**
+ * Formats a field trip's travel date into the given format.
+ */
+function getFormattedRequestTravelDate (
+  request,
+  dateFormat = OTP_API_DATE_FORMAT
+) {
+  return moment(request.travelDate).format(dateFormat)
+}
+
+// These can be overridden in the field trip module config.
+const defaultFieldTripModeCapacities = {
+  'TRAM': 80,
+  'SUBWAY': 120,
+  'RAIL': 80,
+  'BUS': 40,
+  'FERRY': 100,
+  'CABLE_CAR': 20,
+  'GONDOLA': 15,
+  'FUNICULAR': 20
+}
+const unknownModeCapacity = 15
+
+/**
+ * Calculates the mode capacity based on the field trip module mode capacities
+ * (if it exists) or from the above default lookup of mode capacities or if
+ * given an unknown mode, then the unknownModeCapacity is returned.
+ *
+ * @param {Object} config the app-wide config
+ * @param {string} mode the OTP mode
+ */
+function getFieldTripGroupCapacityForMode (fieldTripModuleConfig, mode) {
+  const configModeCapacities = fieldTripModuleConfig?.modeCapacities
+  return (configModeCapacities && configModeCapacities[mode]) ||
+      defaultFieldTripModeCapacities[mode] ||
+      unknownModeCapacity
+}
+
+/**
+ * Dispatches the appropriate cleanup actions after making requests to find the
+ * itineraries for an inbound or outbound field trip journey.
+ */
 function doFieldTripPlanRequestCleanup (dispatch, searchId) {
   // set pending searches to 0 to indicate searching is finished
   dispatch(setPendingRequests({ searchId, pending: 0 }))
