+++ lib/util/call-taker.js
@@ -1,6 +1,5 @@
 import {getRoutingParams} from '@opentripplanner/core-utils/lib/query'
 import {randId} from '@opentripplanner/core-utils/lib/storage'
-import { OTP_API_DATE_FORMAT } from '@opentripplanner/core-utils/lib/time'
 import moment from 'moment'
 import { createSelector } from 'reselect'
 
@@ -217,42 +216,6 @@ export function getTripFromRequest (request, outbound = false) {
   })
   return trip
 }
-
-export function getFormattedRequestTravelDate (
-  request,
-  dateFormat = OTP_API_DATE_FORMAT
-) {
-  return moment(request.travelDate).format(dateFormat)
-}
-
-const defaultFieldTripModeCapacities = {
-  'TRAM': 80,
-  'SUBWAY': 120,
-  'RAIL': 80,
-  'BUS': 40,
-  'FERRY': 100,
-  'CABLE_CAR': 20,
-  'GONDOLA': 10,
-  'FUNICULAR': 20
-}
-const unknownModeCapacity = 10
-
-/**
- * Calculates the mode capacity based on the field trip module mode capacities
- * (if it exists) or from the above default lookup of mode capacities or if
- * given an unknown mode, then the unknownModeCapacity is returned.
- *
- * @param {Object} config the app-wide config
- * @param {string} mode the OTP mode
- */
-export const getFieldTripGroupCapacityForMode = createSelector(
-  fieldTripModuleConfig => fieldTripModuleConfig?.modeCapacities,
-  (fieldTripModuleConfig, mode) => mode,
-  (configModeCapacities, mode) => (configModeCapacities &&
-    configModeCapacities[mode]) ||
-      defaultFieldTripModeCapacities[mode] ||
-      unknownModeCapacity
-)
 
 /**
  * LZW-compress a string
