+++ lib/util/state.js
@@ -93,7 +93,7 @@ export function getResponsesWithErrors (state) {
     response.forEach(res => {
       if (res) {
         if (res.error) {
-          let msg = res.error.msg || 'An error occured while planning a trip'
+          let msg = res.error.msg || 'An error occurred while planning a trip'
           // include the modes if doing batch routing
           if (showModes && res.requestParameters?.mode) {
             const mode = humanReadableMode(res.requestParameters.mode)
